fix(railpack): use ShellQuote to preserve env vars in command
