+++ src/pkg/clouds/do/appPlatform/setup_test.go
@@ -6,38 +6,6 @@ import (
 	"github.com/digitalocean/godo"
 )
 
-func TestShellQuote(t *testing.T) {
-	// Given
-	tests := []struct {
-		input    []string
-		expected string
-	}{
-		{
-			input:    []string{"true"},
-			expected: `"true"`,
-		},
-		{
-			input:    []string{"echo", "hello world"},
-			expected: `"echo" "hello world"`,
-		},
-		{
-			input:    []string{"echo", "hello", "world"},
-			expected: `"echo" "hello" "world"`,
-		},
-		{
-			input:    []string{"echo", `hello"world`},
-			expected: `"echo" "hello\"world"`,
-		},
-	}
-
-	for _, test := range tests {
-		actual := shellQuote(test.input...)
-		if actual != test.expected {
-			t.Errorf("Expected `%s` but got: `%s`", test.expected, actual)
-		}
-	}
-}
-
 func Test_getImageSourceSpec(t *testing.T) {
 	// Given
 	tests := []struct {
